move not-found.component.ts to common
