+++ src/app/common/index.ts
@@ -1,2 +1,3 @@
 export * from './ellipsis-title.directive';
 export * from './window-resize-tracker.service';
+export * from './not-found';
